move CPE related visitors in same file
