+++ cosmo/visitors/router_bgpcpe.py
@@ -2,12 +2,14 @@ from abc import ABCMeta, abstractmethod
 from typing import List, NoReturn, TypeGuard
 
 from multimethod import multimethod as singledispatchmethod
-from ipaddress import IPv4Interface, IPv6Interface
+from ipaddress import IPv4Interface, IPv6Interface, IPv6Network, IPv4Network
 
 from cosmo.common import head, CosmoOutputType, InterfaceSerializationError
 from cosmo.config.cosmo_config import CosmoConfig
-from cosmo.visitors.cpe_router import CpeRouterExporterVisitor, CpeRouterIPVisitor
-from cosmo.visitors.abc import AbstractRouterExporterVisitor
+from cosmo.visitors.abc import (
+    AbstractRouterExporterVisitor,
+    AbstractNoopNetboxTypesVisitor,
+)
 from cosmo.features import features
 from cosmo.log import warn
 from cosmo.manufacturers import ManufacturerFactoryFromDevice
@@ -17,7 +19,60 @@ from cosmo.netbox_types import (
     DeviceType,
     VRFType,
     AbstractNetboxType,
+    IPAddressType,
 )
+
+
+class CpeRouterIPVisitor(AbstractNoopNetboxTypesVisitor):
+    def __init__(self, ip_networks) -> None:
+        super().__init__()
+
+        self.ip_networks = ip_networks
+
+    @singledispatchmethod
+    def accept(self, o):
+        return super().accept(o)
+
+    @accept.register
+    def _(self, o: IPAddressType):
+        ipo = o.getIPInterfaceObject()
+
+        for ipn in self.ip_networks:
+            if o.getIPInterfaceObject() in ipn:
+                return ipo
+
+        return None
+
+
+class CpeRouterExporterVisitor(AbstractNoopNetboxTypesVisitor):
+    """
+    This visitor creates a list of networks which are exported from the router
+    via unnumbered bgp. We allow all configured IP networks on a CPE to be
+    exported. By definition the primary IP is our management IP and this IP
+    should not be allowed to be exported via BGP from the router.
+    """
+
+    def __init__(self, forbidden_networks: list[IPv6Network | IPv4Network]):
+        self.forbidden_networks = forbidden_networks
+
+    @singledispatchmethod
+    def accept(self, o):
+        return super().accept(o)
+
+    @accept.register
+    def _(self, o: IPAddressType):
+        primary_ip4 = o.getParent(DeviceType)["primary_ip4"]
+        if primary_ip4 and primary_ip4.getIPAddress() == o.getIPAddress():
+            return  # skip, they're not allowed to export their mgmt addr
+        ip_interface = o.getIPInterfaceObject()
+        if any(
+            map(
+                lambda forbidden_network: ip_interface in forbidden_network,
+                self.forbidden_networks,
+            )
+        ):
+            return  # skip, they're not allowed to export our transfer nets
+        return type(ip_interface), ip_interface.network.with_prefixlen
 
 
 class AbstractBgpCpeExporter(metaclass=ABCMeta):
